+++ R/class_crew.R
@@ -316,14 +316,7 @@ database_crew <- function(path_store) {
 }
 
 validate_crew_controller <- function(controller) {
-  tar_assert_inherits(
-    x = controller,
-    class = c("crew_class_controller_group", "crew_class_controller"),
-    msg = paste(
-      "controller for tar_make() must be a valid",
-      "object of class \"crew_class_controller\" from the",
-      "{crew} R package."
-    )
-  )
+  tar_assert_envir(controller, msg = "invalid crew controller")
+  tar_assert_function(controller$validate, msg = "invalid crew controller")
   controller$validate()
 }