+++ inst/pipelines/use_targets.R
@@ -1,16 +1,21 @@
 # Created by use_targets().
 # Follow the comments below to fill in this target script.
 # Then follow the manual to check and run the pipeline:
-#   https://books.ropensci.org/targets/walkthrough.html#inspect-the-pipeline # nolint
+#   https://books.ropensci.org/targets/walkthrough.html#inspect-the-pipeline
 
 # Load packages required to define the pipeline:
 library(targets)
-# library(tarchetypes) # Load other packages as needed. # nolint
+# library(tarchetypes) # Load other packages as needed.
 
 # Set target options:
 tar_option_set(
   packages = c("tibble"), # packages that your targets need to run
   format = "rds" # default storage format
+  #
+  # For distributed computing in tar_make(), supply a crew controller
+  #   as discussed at https://books.ropensci.org/targets/crew.html
+  # controller = crew::crew_controller_local(workers = 2)
+  #
   # Set other options as needed.
 )
 
@@ -22,14 +27,14 @@ FUTURE
 
 # Run the R scripts in the R/ folder with your custom functions:
 tar_source()
-# source("other_functions.R") # Source other scripts as needed. # nolint
+# source("other_functions.R") # Source other scripts as needed.
 
 # Replace the target list below with your own:
 list(
   tar_target(
     name = data,
     command = tibble(x = rnorm(100), y = rnorm(100))
-#   format = "feather" # efficient storage of large data frames # nolint
+    # format = "feather" # efficient storage for large data frames
   ),
   tar_target(
     name = model,