+++ R/tar_config_set.R
@@ -45,6 +45,8 @@
 #'   ignored if `tar_option_get("controller")` is `NULL`.
 #'   Independent from the `garbage_collection` argument of [tar_target()],
 #'   which controls garbage collection on the worker.
+#' @param level_separation Argument of [tar_visnetwork()] and [tar_glimpse()]
+#'   to control the space between hierarchical levels.
 #' @param reporter_make Character of length 1, `reporter` argument to
 #'   [tar_make()] and related functions that run the pipeline.
 #'   If the argument `NULL`, the setting is not modified.
@@ -132,6 +134,7 @@
 tar_config_set <- function(
   inherits = NULL,
   garbage_collection = NULL,
+  level_separation = NULL,
   reporter_make = NULL,
   reporter_outdated = NULL,
   script = NULL,
@@ -150,6 +153,7 @@ tar_config_set <- function(
   tar_assert_scalar(project)
   tar_config_assert_inherits(inherits)
   tar_config_assert_garbage_collection(garbage_collection)
+  tar_config_assert_level_separation(level_separation)
   tar_config_assert_reporter_make(reporter_make)
   tar_config_assert_reporter_outdated(reporter_outdated)
   tar_config_assert_script(script)
@@ -164,6 +168,8 @@ tar_config_set <- function(
   yaml[[project]]$inherits <- inherits %|||% yaml[[project]]$inherits
   yaml[[project]]$garbage_collection <- garbage_collection %|||%
     yaml[[project]]$garbage_collection
+  yaml[[project]]$level_separation <- level_separation %|||%
+    yaml[[project]]$level_separation
   yaml[[project]]$reporter_make <- reporter_make %|||%
     yaml[[project]]$reporter_make
   yaml[[project]]$reporter_outdated <- reporter_outdated %|||%
@@ -204,6 +210,15 @@ tar_config_assert_garbage_collection <- function(garbage_collection) {
   tar_assert_lgl(garbage_collection)
   tar_assert_scalar(garbage_collection)
   tar_assert_none_na(garbage_collection)
+}
+
+tar_config_assert_level_separation <- function(level_separation) {
+  if (is.null(level_separation)) {
+    return()
+  }
+  tar_assert_scalar(level_separation)
+  tar_assert_dbl(level_separation)
+  tar_assert_none_na(level_separation)
 }
 
 tar_config_assert_reporter_make <- function(reporter_make) {