+++ R/tar_visnetwork.R
@@ -36,7 +36,7 @@ tar_visnetwork <- function(
   exclude = ".Random.seed",
   outdated = TRUE,
   label = NULL,
-  level_separation = NULL,
+  level_separation = targets::tar_config_get("level_separation"),
   degree_from = 1L,
   degree_to = 1L,
   zoom_speed = 1,