+++ NEWS.md
@@ -1,8 +1,14 @@
 # targets 1.1.3.9000 (development)
 
+## `crew` integration
+
 * Do not assume S3 classes when validating `crew` controllers.
 * Suggest a crew controller in the `_targets.R` file from `use_targets()`.
 * Make `tar_crew()` compatible with `crew` >= 0.2.1.9011.
+
+## Other improvements
+
+* Set default `level_separation` arguments with `tar_config_set()` (#1085, @Moohan).
 
 # targets 1.1.3